Replace index.tsx with newLayout
